refactor(mb-shared): extraire dateSchema et dateTruncSchema dans dates.ts
Les deux schémas étaient définis dans valeurAvancement.ts mais utilisés aussi
par objectifIndicateurIndividu. Déplacés dans un module neutre @pilote/mb-shared/dates.
Tous les consommateurs importent désormais directement depuis ce module.
